+++ test/runtime/samples/raw-mustache-as-root/main.svelte
@@ -1,8 +1,8 @@
 <script>
-	import Component from './Component.svelte';
+	import RawMustache from './RawMustache.svelte';
 
 	let content1 = `<p>First line</p>`;
-	let content2 = `<p>Another first line</p>`
+	let content2 = `<p>Another first line</p>`;
 
 	let show = false;
 	$: content = show ? content1 : content2;
@@ -12,6 +12,6 @@
 	Switch
 </button>
 
-<Component>
-	{@html content}
-</Component>
+<RawMustache {content} />
+
+<p>This line should be last.</p>